+++ PaxstoreOpenApi/Paxstore/OpenApi/Model/App/PagedApp.cs
@@ -10,7 +10,7 @@ namespace Paxstore.OpenApi.Model
         public long ID { get; set; }
         public string Name { get; set; }                // 应用名字
         public string PackageName { get; set; }         // 应用包名,唯一
-        public string status { get; set; }              // 应用状态
+        public string Status { get; set; }              // 应用状态
         public string OsType { get; set; }              // 操作系统
         public Nullable<bool> SpecificReseller { get; set; }   // 是否定向发布至代理商
         public Nullable<int> ChargeType { get; set; }